--- src/main/java/org/sopt/config/auth/LoginMember.java
@@ -1,12 +1,0 @@
-package org.sopt.config.auth;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LoginMember {
-}
-
