--- src/main/java/org/sopt/diary/api/dto/response/DiaryCratedResponse.java
@@ -1,4 +1,0 @@
-package org.sopt.diary.api.dto.response;
-
-public record DiaryCratedResponse(long diaryId) {
-}
