--- src/main/java/org/sopt/diary/api/dto/response/DiaryListResponse.java
@@ -1,6 +1,0 @@
-package org.sopt.diary.api.dto.response;
-
-import java.util.List;
-
-public record DiaryListResponse(List<DiaryResponse> diaries) {
-}
