+++ src/main/java/org/sopt/diary/config/auth/LoginMember.java
@@ -1,0 +1,12 @@
+package org.sopt.diary.config.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginMember {
+}
+
