+++ src/main/java/org/sopt/diary/diaries/api/dto/response/DiaryCratedResponse.java
@@ -1,0 +1,4 @@
+package org.sopt.diary.diaries.api.dto.response;
+
+public record DiaryCratedResponse(long diaryId) {
+}
