+++ src/main/java/org/sopt/diary/diaries/api/dto/response/DiaryListResponse.java
@@ -1,0 +1,6 @@
+package org.sopt.diary.diaries.api.dto.response;
+
+import java.util.List;
+
+public record DiaryListResponse(List<DiaryResponse> diaries) {
+}
