+++ src/main/java/org/sopt/diary/diaries/api/dto/response/DiaryResponse.java
@@ -1,0 +1,10 @@
+package org.sopt.diary.diaries.api.dto.response;
+
+import org.sopt.diary.diaries.domain.Diary;
+
+public record DiaryResponse(long id, String title) {
+
+    public DiaryResponse(Diary diary) {
+        this(diary.getId(), diary.getTitle());
+    }
+}
