--- src/main/java/org/sopt/diary/domain/Diary.java
@@ -1,62 +1,0 @@
-package org.sopt.diary.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
-import org.sopt.config.domain.BaseEntity;
-import org.sopt.member.domain.Member;
-
-@Entity
-public class Diary extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 30)
-    private String title;
-
-    @Column(nullable = false, length = 100)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Member member;
-
-    protected Diary() {
-    }
-
-    public Diary(Long id, String title, String content, Member member) {
-        this.id = id;
-        this.title = title;
-        this.content = content;
-        this.member = member;
-    }
-
-    public Diary(String title, String content, Member member) {
-        this(null, title, content, member);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    public void update(String title, String content) {
-        this.title = title;
-        this.content = content;
-    }
-}
