+++ src/main/java/org/sopt/diary/member/api/dto/LoginRequest.java
@@ -1,0 +1,4 @@
+package org.sopt.diary.member.api.dto;
+
+public record LoginRequest(String username, String password) {
+}
