+++ src/main/java/org/sopt/diary/member/api/dto/LoginResponse.java
@@ -1,0 +1,10 @@
+package org.sopt.diary.member.api.dto;
+
+import org.sopt.diary.member.domain.Member;
+
+public record LoginResponse(long memberId) {
+
+    public LoginResponse(Member member) {
+        this(member.getId());
+    }
+}
