--- src/main/java/org/sopt/member/api/dto/LoginRequest.java
@@ -1,4 +1,0 @@
-package org.sopt.member.api.dto;
-
-public record LoginRequest(String username, String password) {
-}
