--- src/main/java/org/sopt/member/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package org.sopt.member.repository;
-
-import java.util.Optional;
-import org.sopt.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByUsername(String username);
-}
